Functional, not optimized.
This script creates a series of CW Dashboards with the common EBS volume metrics that show Latency, Queue Depth, and Impaired Volume status.

The CW Dashboards are created based on the Tag Name and Tag Value specified on the command line.

At the end, of the script, any dashboards that are stale for that tag name will be deleted unless the --no-cleanup argument is added.

Use the --tag-name to specify the tag to create dashboards for.

Each dashboard will have the name "EBS_<tag_name>_<tag_value>_<#>" where # is an incrementing number based on the DASHBOARDS_METRICS_LIMIT constant.

Todo:
*